refactored data processing to reduce ram usage
